build(test): enable MVS builds of test/ sources

+++ test/mvs/tstarext.c
@@ -1,5 +1,5 @@
 /* ------------------------------------------------------------------ */
-/*  test_arith_extended.c - WP-21b Phase B tests for the new public    */
+/*  tstarext.c - WP-21b Phase B tests for the new public    */
 /*  IRXARITH string-oriented helpers: trunc, format, from_digits,      */
 /*  to_digits. These are direct C-API tests; they do not drive the     */
 /*  parser.                                                            */
@@ -13,7 +13,6 @@
 /*  (c) 2026 mvslovers - REXX/370 Project                            */
 /* ------------------------------------------------------------------ */
 
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -55,7 +54,7 @@ static int tests_failed = 0;
 static Lstr make_lstr(const char *cstr)
 {
     Lstr s;
-    s.pstr = (unsigned char *)(uintptr_t)cstr;
+    s.pstr = (unsigned char *)(unsigned long)cstr;
     s.len = strlen(cstr);
     s.maxlen = s.len;
     s.type = LSTRING_TY;
